% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{aggregate_population_metrics}
\alias{aggregate_population_metrics}
\title{Aggregate Metrics from Population}
\usage{
aggregate_population_metrics(
  population,
  train_data_obj,
  valid_data_obj,
  aggregation = "mean"
)
}
\arguments{
\item{population}{Population object}

\item{train_data_obj}{Training data as Gpredomics Data object (or NULL)}

\item{valid_data_obj}{validuation data as Gpredomics Data object (or NULL)}

\item{aggregation}{Aggregation method: "mean" or "median"}
}
\value{
Named list of aggregated metrics with train_ and valid_ prefixes
}
\description{
Internal function to compute and aggregate metrics from individual models in a population.
}
\keyword{internal}
