% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{analyzeAttributeEvolution}
\alias{analyzeAttributeEvolution}
\title{Analyze evolution of model attributes over generations}
\usage{
analyzeAttributeEvolution(
  exp,
  attributes = c("auc", "fit", "k", "epoch"),
  best_model = TRUE,
  plot = TRUE
)
}
\arguments{
\item{exp}{A list representing the experiment data, containing populations over generations.}

\item{attributes}{A vector of attribute names to analyze (default: c("auc", "fit", "k", "epoch")).}

\item{best_model}{Logical; should the analyses focus on the best model? (default: TRUE).}

\item{plot}{Logical; should a plot be generated? (default: TRUE).}
}
\value{
A dataframe containing the evolution of attributes over generations, or a plot if \code{plot = TRUE}.
}
\description{
Extracts and visualizes how selected model attributes change over generations in an experiment.
}
