% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.X_y}
\alias{check.X_y}
\title{Validate dataset and response vector consistency}
\usage{
check.X_y(X, y)
}
\arguments{
\item{X}{A data matrix or dataframe.}

\item{y}{A response vector.}
}
\value{
None; stops execution with an error if validation fails.
}
\description{
Checks if X (dataset) and y (response vector) have coherent dimensions.
}
