% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.Individual}
\alias{coef.Individual}
\title{Extracts coefficients from an Individual object}
\usage{
\method{coef}{Individual}(object, ...)
}
\arguments{
\item{object}{An object of class \code{Individual} (Rust-backed external pointer).}
}
\value{
A named vector containing the coefficients of the Individual.
}
\description{
This method retrieves the coefficients of an Individual object.
}
