% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{compute_auc_individual}
\alias{compute_auc_individual}
\title{Compute AUC for an Individual}
\usage{
compute_auc_individual(individual, data)
}
\arguments{
\item{individual}{An Individual object}

\item{data}{A Data object}
}
\description{
Compute AUC for an Individual
}
