% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dim.Data}
\alias{dim.Data}
\title{Gets the dimensions of a Data object}
\usage{
\method{dim}{Data}(data)
}
\arguments{
\item{data}{An object of class \code{Data} (Rust-backed external pointer).}
}
\value{
A numeric vector containing the dimensions of the Data object.
}
\description{
This method returns the dimensions of a Data object.
}
