% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{evaluate_individual}
\alias{evaluate_individual}
\title{Evaluate an Individual}
\usage{
evaluate_individual(individual, data)
}
\arguments{
\item{individual}{An Individual object}

\item{data}{A Data object (assuming Data object is defined similarly)}
}
\description{
Evaluate an Individual
}
