% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpredomics_caret.R
\name{getModelInfo_gpredomics}
\alias{getModelInfo_gpredomics}
\title{Get Gpredomics Model Specification for Caret}
\usage{
getModelInfo_gpredomics()
}
\value{
List with model specification compatible with caret
}
\description{
Returns the model specification list that can be passed directly to caret::train().
This is the recommended way to use gpredomics with caret.
}
\examples{
\dontrun{
library(caret)
library(gpredomicsR)

# Get model specification
gp_model <- getModelInfo_gpredomics()

# Use directly with train()
ctrl <- trainControl(method = "cv", number = 5)
model <- train(Species ~ ., data = iris_binary, 
               method = gp_model, trControl = ctrl)
}
}
