% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{getPopulationAttribute}
\alias{getPopulationAttribute}
\title{Retrieve Specific Attribute for Each Individual in a Population}
\usage{
getPopulationAttribute(pop, attribute = "fit")
}
\arguments{
\item{pop}{A list of lists, where each inner list represents an individual and should contain named attributes.}

\item{attribute}{A character string specifying the name of the attribute to extract from each individual. The default is "fit".}
}
\value{
A vector containing the values of the specified attribute for each individual in the population. If an individual does not have the attribute, the corresponding entry in the vector will be \code{NA}.
}
\description{
This function iterates through a list representing a population and extracts a specified attribute from each individual. If the attribute is not present, \code{NA} is assigned for that individual.
}
\examples{
# Assuming a list of individuals with attributes
individuals <- list(
  list(fit = 0.95, auc = 0.90),
  list(fit = 0.89),
  list(auc = 0.92)  # This individual does not have 'fit'
)
# Extract 'fit' attribute
fits <- getPopulationAttribute(individuals, attribute = "fit")
print(fits)  # Prints: c(0.95, 0.89, NA)

}
