% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTheBestIndividual}
\alias{getTheBestIndividual}
\title{Retrieve the Best Individual from a Population}
\usage{
getTheBestIndividual(pop, evalToFit = "fit")
}
\arguments{
\item{pop}{A list representing a population of models.}

\item{evalToFit}{A character string indicating the attribute used for sorting (default: "fit").}
}
\value{
The best individual from the population or NULL if the population is empty or all values are NA.
}
\description{
Sorts a population based on a given evaluation criterion and returns the best individual.
}
