% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{gp_set_log_level}
\alias{gp_set_log_level}
\title{Set the R-side log level}
\usage{
gp_set_log_level(level = c("info", "warn", "error", "debug", "none"))
}
\arguments{
\item{level}{one of c("none","error","warn","info","debug")}
}
\description{
Set the R-side log level
}
\keyword{internal}
