% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpredomics_caret.R
\name{gpredomics_info}
\alias{gpredomics_info}
\title{Get Gpredomics Model Info for Caret}
\usage{
gpredomics_info()
}
\value{
List with model metadata
}
\description{
Returns model information for caret registration.
}
