% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpredomics_caret.R
\name{gpredomics_predict}
\alias{gpredomics_predict}
\title{Predict from Gpredomics Model via Caret}
\usage{
gpredomics_predict(modelFit, newdata, preProc = NULL, submodels = NULL, ...)
}
\arguments{
\item{modelFit}{Model object returned by gpredomics_train()}

\item{newdata}{New data for prediction (data.frame or matrix)}

\item{preProc}{Pre-processing object (currently not used)}

\item{submodels}{List of submodels (currently not used)}

\item{...}{Additional arguments (not used)}
}
\value{
Predictions (factor for classification with levels 0 and 1)
}
\description{
Prediction function compatible with caret's train() interface.
This function is called internally by caret for generating predictions.
}
