% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpredomics_caret.R
\name{gpredomics_train}
\alias{gpredomics_train}
\title{Train Gpredomics Model via Caret}
\usage{
gpredomics_train(x, y, wts = NULL, param, lev, last, weights, classProbs, ...)
}
\arguments{
\item{x}{Feature matrix (samples in rows, features in columns)}

\item{y}{Response vector (binary: 0/1)}

\item{wts}{Sample weights (currently not used)}

\item{param}{Data.frame with a single row of hyperparameters}

\item{lev}{Class levels}

\item{last}{Logical, whether this is the final model}

\item{weights}{Sample weights (currently not used)}

\item{classProbs}{Logical, whether to compute class probabilities}

\item{...}{Additional arguments}
}
\value{
A list containing the trained experiment and metadata
}
\description{
Training function compatible with caret's train() interface.
This function is called internally by caret during hyperparameter tuning.
}
