% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isExperiment}
\alias{isExperiment}
\title{Check if an Object is an Experiment}
\usage{
isExperiment(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
TRUE if the object is an experiment, FALSE otherwise.
}
\description{
This function evaluates whether an object is a valid experiment.
It ensures that the object is a list and contains essential attributes.
}
