% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isModel}
\alias{isModel}
\title{Evaluates whether an object is a model}
\usage{
isModel(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
TRUE if the object is a model, FALSE otherwise.
}
\description{
Checks whether an object conforms to the expected model structure.
}
