% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isModelCollection}
\alias{isModelCollection}
\title{Evaluates whether an object is a model collection object}
\usage{
isModelCollection(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
TRUE if the object is a model collection, FALSE otherwise.
}
\description{
Checks whether an object is a collection of model populations.
}
