% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isPopulation}
\alias{isPopulation}
\title{Evaluates whether an object is a population of models}
\usage{
isPopulation(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
TRUE if the object is a population of models, FALSE otherwise.
}
\description{
Checks whether an object is a list of valid model objects.
}
