% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length.Experiment}
\alias{length.Experiment}
\title{Gets the number of generations in an Experiment}
\usage{
\method{length}{Experiment}(exp)
}
\arguments{
\item{exp}{An object of class \code{Experiment} (Rust-backed external pointer).}
}
\value{
An integer representing the number of generations in the Experiment.
}
\description{
This method returns the number of generations in an Experiment object.
}
