% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length.Jury}
\alias{length.Jury}
\title{Gets the Jury population size}
\usage{
\method{length}{Jury}(jury)
}
\arguments{
\item{jury}{An object of class \code{Jury} (Rust-backed external pointer).}
}
\value{
An integer representing the size of the Jury population.
}
\description{
This method returns the size of the Jury population.
}
