% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length.Population}
\alias{length.Population}
\title{Gets the Population size}
\usage{
\method{length}{Population}(pop)
}
\arguments{
\item{pop}{An object of class \code{Population} (Rust-backed external pointer).}
}
\value{
An integer representing the size of the Population.
}
\description{
This method returns the size of a Population object.
}
