% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listOfModelsToListOfDenseVec}
\alias{listOfModelsToListOfDenseVec}
\title{Builds a list of dense vector coefficients from a list of models}
\usage{
listOfModelsToListOfDenseVec(X, y, list.models)
}
\arguments{
\item{X}{A dataset (matrix or dataframe).}

\item{y}{A vector of labels.}

\item{list.models}{A list of model objects.}
}
\value{
A list of dense vectors of coefficients.
}
\description{
Builds a list of dense vector coefficients from a list of models
}
