% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modelToDenseVec}
\alias{modelToDenseVec}
\title{Transform the model object into a dense coefficient vector}
\usage{
modelToDenseVec(natts, mod)
}
\arguments{
\item{natts}{The number of attributes (features).}

\item{mod}{A model object.}
}
\value{
A dense vector representing the model coefficients.
}
\description{
Converts a model object into a dense (long) format vector.
}
