% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{names.Data}
\alias{names.Data}
\title{Gets the names of fields in a Data object}
\usage{
\method{names}{Data}(data)
}
\arguments{
\item{data}{An object of class \code{Data} (Rust-backed external pointer).}
}
\value{
A character vector containing the names of the fields in the Data object.
}
\description{
This method returns the names of the fields in a Data object.
}
