% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{parseExperiment}
\alias{parseExperiment}
\title{parseExperiment}
\usage{
parseExperiment(exp)
}
\arguments{
\item{exp}{The experiment object}
}
\value{
A list of generations, each generation containing a list of individuals
}
\description{
Parse the experiment object to extract the individuals
}
