% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.Experiment}
\alias{plot.Experiment}
\title{Plots Training History from an Experiment Object}
\usage{
\method{plot}{Experiment}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Experiment} (Rust-backed external pointer).}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A ggplot object representing the training history.
}
\description{
This method generates a plot of the training history from an Experiment object.
}
