% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{plot_genealogy}
\alias{plot_genealogy}
\title{Plot a precomputed genealogy (Graph / Tree view)}
\usage{
plot_genealogy(
  ind,
  exp,
  max_depth = 3,
  layout = "sugiyama",
  node_size_base = 2,
  text_repel = TRUE,
  node_vars = list(color = "auc", size = "k", label = "label")
)
}
\arguments{
\item{ind}{Individual to plot genealogy for.}

\item{max_depth}{Maximum depth of the genealogy to display.}

\item{layout}{Graph layout: "tree", "kk", "fr", "sugiyama".}

\item{node_size_base}{Base size of the nodes.}

\item{text_repel}{Use ggrepel for labels.}

\item{node_vars}{List or column names: list(color="auc", size="k", label="label").}

\item{experiment}{Experiment object.}
}
\value{
A ggplot/ggraph object.
}
\description{
Plot a precomputed genealogy (Graph / Tree view)
}
