% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{populationGet_X}
\alias{populationGet_X}
\title{Get the best model from a classifier result}
\usage{
populationGet_X(element2get, toVec = TRUE, na.rm = TRUE)
}
\arguments{
\item{element2get}{A character string specifying the attribute to extract.}

\item{toVec}{Logical; should the results be unlisted into a vector? (default: TRUE)}

\item{na.rm}{Logical; should NA values be removed? (default: TRUE)}
}
\value{
A vector or list of extracted attributes, depending on \code{toVec}.
}
\description{
Extracts a given attribute from a population of predomics objects.
}
