% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.Jury}
\alias{predict.Jury}
\title{Predicts classes or scores for an Jury object}
\usage{
\method{predict}{Jury}(object, new_data, type = c("class", "score"), ...)
}
\arguments{
\item{object}{An object of class \code{Jury}.}

\item{new_data}{A Data object for prediction.}

\item{type}{Character: "class" (default) or "score" (raw decision values).}
}
\description{
Predicts classes or scores for an Jury object
}
