% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{print.gpredomics_grid_search}
\alias{print.gpredomics_grid_search}
\title{Print method for grid search results}
\usage{
\method{print}{gpredomics_grid_search}(x, ...)
}
\arguments{
\item{x}{A gpredomics_grid_search object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for grid search results
}
