% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{printExperiment}
\alias{printExperiment}
\title{Print Information about an Experiment Object}
\usage{
printExperiment(obj, indent = "\\t--- ")
}
\arguments{
\item{obj}{An experiment object created by \code{runExperiment()}.}

\item{indent}{A string for indentation to structure the output.}
}
\value{
None. The function prints experiment details to the console.
}
\description{
This function prints details about an experiment object, including execution time,
parameters, dataset, and model collection.
}
