% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{printy}
\alias{printy}
\title{Print Summary of Predomics Object}
\usage{
printy(obj)
}
\arguments{
\item{obj}{An object that can be of type model, population, classifier,
experiment, or model collection.}
}
\value{
None. The function prints the summary of the object directly to the console.
}
\description{
This function prints a summary of a given object, identifying its type
(model, population, classifier, experiment, or model collection) and calling
the appropriate print function to display relevant information about the object.
}
\details{
The function checks the type of the provided object using \code{isModel},
\code{isPopulation}, \code{isClf}, \code{isExperiment}, and \code{isModelCollection} functions.
Based on the object type, it prints a summary:
\itemize{
\item \strong{Model}: Calls \code{printModel} with a detailed description of the model.
\item \strong{Population}: Calls \code{printPopulation}, showing a summary of the population of models.
\item \strong{Model Collection}: Calls \code{printModelCollection} to show a summary of a collection of models.
}

If the object type is not recognized, an error message is printed.
}
\examples{
\dontrun{
# Assuming 'model', 'population', 'classifier', 'experiment', and 'model_collection' are valid objects
printy(model)
printy(population)
printy(model_collection)
}

}
\author{
Edi Prifti (IRD)
}
