% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectBestPopulation}
\alias{selectBestPopulation}
\title{Select the Top Significant Best Models from a Population}
\usage{
selectBestPopulation(pop, score = "fit", p = 0.05, k_penalty = 0, k_max = 0)
}
\arguments{
\item{pop}{A list of model objects (i.e., a population of models).}

\item{score}{The attribute of the model used for evaluation (default: "fit").}

\item{p}{The p-value threshold for statistical significance (default: 0.05).}

\item{k_penalty}{A penalty applied to the score based on the k-sparsity value (default: 0).}

\item{k_max}{A threshold for sparsity selection (default: 0, meaning no sparsity filtering).}
}
\value{
A filtered sub-population of models or \code{NULL} if no models meet the criteria.
}
\description{
Selects the best part of a population where models are significantly \strong{not different} from the best model.
}
