% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortPopulation}
\alias{sortPopulation}
\title{Sort a Population Based on an Evaluation Criterion}
\usage{
sortPopulation(pop, evalToOrder = "fit", decreasing = TRUE)
}
\arguments{
\item{pop}{A list representing a population of models.}

\item{evalToOrder}{A character string indicating the attribute used for sorting (default: "fit").}

\item{decreasing}{Logical; should sorting be in decreasing order? (default: TRUE).}
}
\value{
A sorted list of models based on the specified attribute.
}
\description{
Sorts a population of models based on a specified attribute.
}
