% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.Experiment}
\alias{[.Experiment}
\title{Extracts Generations from an Experiment object}
\usage{
\method{[}{Experiment}(exp, generation, fold, on_train = TRUE)
}
\arguments{
\item{exp}{An object of class \code{Experiment} (Rust-backed external pointer).}

\item{generation}{A numeric value specifying the generation number to extract.}

\item{fold}{A numeric value specifying the fold number (for cross-validation).}

\item{on_train}{A logical value indicating whether to access the population fitted on training set (default is
TRUE).}
}
\value{
A Generation object corresponding to the specified generation and fold.
}
\description{
This method allows for indexing into an Experiment object to
retrieve specific Generations based on generation number and fold.
}
