% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.Jury}
\alias{[.Jury}
\title{Extracts parameters from a Jury object}
\usage{
\method{[}{Jury}(jury, accessions)
}
\arguments{
\item{jury}{An object of class \code{Jury} (Rust-backed external pointer).}

\item{accessions}{A numeric vector of indices or a logical vector as a mask.}
}
\value{
A list of Individual objects corresponding to the specified indices or mask.
}
\description{
This method allows for indexing into a Jury object to
retrieve specific Individuals based on numeric indices or logical masks.
}
