% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.Population}
\alias{[.Population}
\title{Extracts Individuals from a Population object}
\usage{
\method{[}{Population}(pop, accessions)
}
\arguments{
\item{pop}{An object of class \code{Population} (Rust-backed external pointer).}

\item{accessions}{A numeric vector of indices or a logical vector as a mask.}
}
\value{
A list of Individual objects corresponding to the specified indices or mask.
}
\description{
This method allows for indexing into a Population object to
retrieve specific Individuals based on numeric indices or logical masks.
}
