% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[.Data}
\alias{[[.Data}
\title{Extracts fields from a Data object}
\usage{
\method{[[}{Data}(data, field)
}
\arguments{
\item{data}{An object of class \code{Data} (Rust-backed external pointer).}

\item{field}{A character string specifying the field name to extract.}
}
\value{
A vector or list containing the requested field values.
}
\description{
This method allows for the extraction of internal Rust-managed
fields from a Data object into an R-compatible format.
}
