% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[.Individual}
\alias{[[.Individual}
\title{Extracts fields from an Individual object}
\usage{
\method{[[}{Individual}(individual, field)
}
\arguments{
\item{individual}{An object of class \code{Individual} (Rust-backed external pointer).}

\item{field}{A character string specifying the field name to extract.}
}
\value{
A vector or list containing the requested field values.
}
\description{
This method allows for the extraction of internal Rust-managed
fields from an Individual object into an R-compatible format.
}
