% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[.Param}
\alias{[[.Param}
\title{Extracts a parameter from a Param object}
\usage{
\method{[[}{Param}(param, parameter)
}
\arguments{
\item{param}{An object of class \code{Param} (Rust-backed external pointer).}

\item{parameter}{A character string specifying the parameter name to extract.}
}
\value{
The value of the specified parameter.
}
\description{
This method allows for the extraction of parameters from a Param object
using the double bracket operator.
}
