% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[.Population}
\alias{[[.Population}
\title{Extracts fields from a Population object}
\usage{
\method{[[}{Population}(pop, field)
}
\arguments{
\item{pop}{An object of class \code{Population} (Rust-backed external pointer).}

\item{field}{A character string specifying the field name to extract from each Individual.}
}
\value{
A vector containing the requested field values from all Individuals in the Population.
}
\description{
This method allows for the extraction of specific fields from all
Individuals within a Population object.
}
