% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[<-.Param}
\alias{[[<-.Param}
\title{Sets a parameter in a Param object}
\usage{
\method{[[}{Param}(param, parameter) <- value
}
\arguments{
\item{param}{An object of class \code{Param} (Rust-backed external pointer).}

\item{parameter}{A character string specifying the parameter name to set.}

\item{value}{The value to assign to the specified parameter.}
}
\value{
The updated Param object.
}
\description{
This method allows for setting parameters in a Param object
using the double bracket assignment operator.
}
