% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Data}
\alias{Data}
\alias{$.Data}
\title{Data}
\format{
An object of class \code{environment} of length 4.
}
\usage{
Data
}
\description{
Gpredomics Data object containing feature matrix and labels.
}
\details{
The Data object stores the feature matrix (X), labels (y), sample names, feature names,
and other metadata needed for machine learning algorithms.
}
\section{Methods}{

\describe{
\item{\code{new()}}{Create a new empty Data object.}
\item{\code{get()}}{Returns an R list with all Data fields (X, y, features, samples, etc.).}
\item{\code{address()}}{Get memory address of this Data object as a string.}
\item{\code{print()}}{Get a formatted string summary of the Data dimensions and content.}
}
}

\keyword{datasets}
