% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Experiment}
\alias{Experiment}
\alias{$.Experiment}
\title{TODO add a load function to load a new Data (and check_compatibility)}
\format{
An object of class \code{environment} of length 12.
}
\usage{
Experiment
}
\description{
TODO add a load function to load a new Data (and check_compatibility)
}
\keyword{datasets}
