% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Param}
\alias{Param}
\alias{$.Param}
\title{Param}
\format{
An object of class \code{environment} of length 7.
}
\usage{
Param
}
\description{
Gpredomics parameter object that stores all algorithm settings.
}
\details{
The Param object contains all configuration settings for running gpredomics algorithms
including genetic algorithm parameters, data parameters, cross-validation settings, etc.
}
\section{Methods}{

\describe{
\item{\code{new()}}{Create a new empty Param object.}
\item{\code{load(file_path)}}{Load a param.yaml file to create a new Param.
\itemize{
\item \code{file_path}: Path to param.yaml file
}
}
\item{\code{get()}}{Returns an R list representing the current state of Param with all settings.}
\item{\code{set(variable, value)}}{Set a numeric parameter by name.
\itemize{
\item \code{variable}: Name of the parameter to set
\item \code{value}: New numeric value for the parameter
}
}
\item{\code{set_string(variable, string)}}{Set a string parameter by name.
\itemize{
\item \code{variable}: Name of the parameter to set
\item \code{string}: New string value for the parameter
}
}
\item{\code{set_bool(variable, value)}}{Set a boolean parameter by name.
\itemize{
\item \code{variable}: Name of the parameter to set
\item \code{value}: New boolean value for the parameter
}
}
\item{\code{address()}}{Get memory address of this Param object as a string.}
}
}

\keyword{datasets}
