% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{RunningFlag}
\alias{RunningFlag}
\alias{$.RunningFlag}
\title{RunningFlag}
\format{
An object of class \code{environment} of length 4.
}
\usage{
RunningFlag
}
\description{
A struct to manage the running flag for controlling algorithm execution.
}
\details{
The RunningFlag object allows you to control the execution of long-running algorithms.
It can be used to stop algorithms gracefully from another thread or after user interruption.
}
\section{Methods}{

\describe{
\item{\code{new()}}{Create a new RunningFlag (initially set to TRUE).}
\item{\code{stop()}}{Set the running flag to FALSE to signal algorithm termination.}
\item{\code{is_running()}}{Check the current value of the running flag (TRUE or FALSE).}
\item{\code{reset()}}{Reset the running flag to TRUE.}
}
}

\keyword{datasets}
